Added max temperature setting

+++ Firmware/core/HAL.h
@@ -6,7 +6,7 @@ namespace HAL {
 void delay(std::uint32_t ms);
 
 namespace Display {
-void set(float backlight_percent, float contrast_percent);
+void set(uint8_t backlight_percent, uint8_t contrast_percent);
 void write(char line1[17], char line2[17]);
 }  // namespace Display
