+++ Firmware/targets/arm/Src/HAL.cpp
@@ -24,9 +24,9 @@ void delay(uint32_t ms) {
 }
 
 namespace Display {
-void set(float backlight_percent, float contrast_percent) {
-    __HAL_TIM_SET_COMPARE(&htim3, TIM_CHANNEL_4, static_cast<uint16_t>(backlight_percent));
-    __HAL_TIM_SET_COMPARE(&htim4, TIM_CHANNEL_4, static_cast<uint16_t>(contrast_percent));
+void set(uint8_t backlight_percent, uint8_t contrast_percent) {
+    __HAL_TIM_SET_COMPARE(&htim3, TIM_CHANNEL_4, backlight_percent);
+    __HAL_TIM_SET_COMPARE(&htim4, TIM_CHANNEL_4, contrast_percent);
 }
 
 void write(char line1[17], char line2[17]) {
